TP03: HTTP y Errores (#108)

+++ src/main/java/com/imb2025/smedico/service/jpa/MedicamentoServiceImpl.java
@@ -38,7 +38,7 @@ public class MedicamentoServiceImpl implements IMedicamentoService{
 		return repoMedic.save(medicamento);	
 	}
 	
-	
+	/*
 	@Override
 	public Medicamento update(Long id, Medicamento medicamento) throws Exception{
 		 Optional<Medicamento> opt = repoMedic.findById(id);
@@ -52,16 +52,57 @@ public class MedicamentoServiceImpl implements IMedicamentoService{
 		    	throw new Exception("El medicamento con ID " + id + " no existe.");
 		    }	
 	}
+	*/
+	
+	
+	@Override
+	public Medicamento update(Long id, Medicamento medicamento) throws Exception {
+	    if (!repoMedic.existsById(id)) {
+	        throw new Exception("El medicamento con ID " + id + " no existe.");
+	    }
 
+	    Medicamento medExistente = repoMedic.findById(id).get();
+
+	    medExistente.setNombre(medicamento.getNombre());
+	    medExistente.setDosisSugerida(medicamento.getDosisSugerida());
+	    medExistente.setPresentacion(medicamento.getPresentacion());
+
+	    return repoMedic.save(medExistente);
+	}
+	
+	@Override
+	public boolean existsById(Long id) {
+		return repoMedic.existsById(id);
+	}
+	
+	
 	@Override
 	public void deleteById(Long id) {
+		if(!repoMedic.existsById(id)) {
+			throw new IllegalArgumentException("El Medicamento con ID: "+ id +" no existe.");
+		}
 		repoMedic.deleteById(id);		
 	}
 	
 	@Override
 	public Medicamento fromDto(MedicamentoRequestDTO dto) throws Exception {
+		if (dto.getNombre() == null || dto.getNombre().isBlank()) {
+            throw new IllegalArgumentException("El nombre no puede estar nulo o vacío");
+        }
+		if (dto.getDosisSugerida() == null || dto.getDosisSugerida().isBlank()) {
+            throw new IllegalArgumentException("El nombre no puede estar nulo o vacío");
+        }
+		if (dto.getPresentacion() == null || dto.getPresentacion().isBlank()) {
+            throw new IllegalArgumentException("El nombre no puede estar nulo o vacío");
+        }
 		
-	    return new Medicamento(dto.getNombre(), dto.getDosisSugerida(), dto.getPresentacion());
+		Medicamento medicamento = new Medicamento();
+		
+		medicamento.setNombre(dto.getNombre());
+		medicamento.setDosisSugerida(dto.getDosisSugerida());
+		medicamento.setPresentacion(dto.getPresentacion());
+	    
+		return medicamento;
 	}
